poc-grok-cli/requirements.txt
click==8.0.1
GitPython==3.1.18